更新Backend和TableModel类，优化日志和车次信息处理
在Backend和TableModel类中进行了一系列的更新和优化。主要改动包括：在Backend.cpp中添加了新的静态成员today_log用于记录当天的日志信息，修改了多个函数以改进车次信息和日志的处理，包括将busNumber的类型从int改为QString，使loadBusInfo函数返回布尔值表示文件加载是否成功，以及在addBusInfo和removeBusInfo函数中将信息保存到两个文件中。同时，也在Backend.h中更新了BusInfo结构和Backend类。在TableModel::updateData函数中，将过滤条件从int类型的车次号改为QString类型的车次号。

+++ Bus/TableModel.cpp
@@ -99,7 +99,7 @@ void TableModel::updateData()
 	QVector<BusInfo> filtered_data;
 	for (auto& bus : all_data) {
 		if (Filter.isEmpty() ||
-			(FilterOption == 0 && bus.busNumber == Filter.toInt()) ||
+			(FilterOption == 0 && bus.busNumber == Filter) ||
 			(FilterOption == 1 && bus.startPoint == Filter) ||
 			(FilterOption == 2 && bus.endPoint == Filter)) {
 			filtered_data.append(bus);
